requirements.txt
odroid_wiringpi
git+https://github.com/unims77/odroid_c1_dht11.git
keyboard~=0.13.5
luma.led_matrix~=3.5.1